WEB-33990: Backpressure for MsbResponderActors
The MsbResponderActor calls a handleRequest method which is implemented
by the service. This service can then choose to either return an object
or a Future. If it returns a Future, the MsbResponderActor will detect
it and wait for its completion, before returning from the
io.github.tcdl.msb.api.ResponderServer.RequestHandler#process method.